+++ application/routes/automation_blueprints.py
@@ -9,7 +9,7 @@
 #           https://github.com/LukedeMunk/zyrax-home-main-controller
 #
 ################################################################################
-from flask import Blueprint, request                                            #Import flask blueprints and requests
+from flask import Blueprint, request, session                                   #Import flask blueprints and requests
 import configuration as c                                                       #Import application configuration variables
 from DeviceManager import DeviceManager                                         #Import device manager
 from server_manager import generate_json_http_response
@@ -25,6 +25,9 @@ automation_bp = Blueprint("automation_blueprints", __name__)
 ################################################################################
 @automation_bp.route("/add_automation", methods=["POST"])
 def add_automation():
+    if "account_id" not in session:
+        return generate_json_http_response(c.HTTP_CODE_UNAUTHORIZED)
+    
     request_data = request.json
     automation_dict = generate_automation_dict_from_request(request_data)
     result = dm.add_automation(automation_dict)
@@ -41,6 +44,9 @@ def add_automation():
 ################################################################################
 @automation_bp.route("/update_automation", methods=["POST"])
 def update_automation():
+    if "account_id" not in session:
+        return generate_json_http_response(c.HTTP_CODE_UNAUTHORIZED)
+    
     request_data = request.json
     automation_dict = generate_automation_dict_from_request(request_data)
     dm.update_automation(int(request_data.get("id")), automation_dict)
@@ -54,6 +60,9 @@ def update_automation():
 ################################################################################
 @automation_bp.route("/delete_automation", methods=["POST"])
 def delete_automation():
+    if "account_id" not in session:
+        return generate_json_http_response(c.HTTP_CODE_UNAUTHORIZED)
+    
     id = int(request.form.get("id"))
     dm.delete_automation(id)
 
@@ -66,6 +75,9 @@ def delete_automation():
 ################################################################################
 @automation_bp.route("/set_automation_enabled", methods=["POST"])
 def set_automation_enabled():
+    if "account_id" not in session:
+        return generate_json_http_response(c.HTTP_CODE_UNAUTHORIZED)
+    
     id = int(request.form.get("id"))
     enabled = int(request.form.get("enabled"))
     dm.set_automation_enabled(id, enabled)
@@ -82,6 +94,9 @@ def set_automation_enabled():
 #
 ################################################################################
 def generate_automation_dict_from_request(request_data):
+    if "account_id" not in session:
+        return generate_json_http_response(c.HTTP_CODE_UNAUTHORIZED)
+    
     automation_dict = {
                     "name": request_data.get("name"),
                     "target_device_ids": request_data.get("target_device_ids"),
@@ -97,7 +112,7 @@ def generate_automation_dict_from_request(request_data):
         if automation_dict["inverted_automation_copy_id"] != -1:
             automation_dict["inverted_action_time"] = request_data.get("inverted_action_time")
 
-    if automation_dict["trigger"] == c.AUTOMATION_TRIGGER_DOOR_SENSOR or automation_dict["trigger"] == c.AUTOMATION_TRIGGER_MOTION_SENSOR:
+    if automation_dict["trigger"] == c.AUTOMATION_TRIGGER_SENSOR:
         automation_dict["time_window_activated"] = int(request_data.get("time_window_activated"))
         automation_dict["activate_during_time_window"] = int(request_data.get("activate_during_time_window"))
 
@@ -105,6 +120,14 @@ def generate_automation_dict_from_request(request_data):
             automation_dict["time_window_start_minutes"] = int(request_data.get("time_window_start_minutes"))
             automation_dict["time_window_end_minutes"] = int(request_data.get("time_window_end_minutes"))
 
+        automation_dict["trigger_device_ids"] = request_data.get("trigger_device_ids")
+        automation_dict["trigger_state"] = int(request_data.get("trigger_state"))
+        automation_dict["delay_minutes"] = int(request_data.get("delay_minutes"))
+
+        if automation_dict["inverted_automation_copy_id"] != -1:
+            automation_dict["inverted_delay_minutes"] = int(request_data.get("inverted_delay_minutes"))
+
+    if automation_dict["trigger"] == c.AUTOMATION_TRIGGER_SWITCH:
         automation_dict["trigger_device_ids"] = request_data.get("trigger_device_ids")
         automation_dict["trigger_state"] = int(request_data.get("trigger_state"))
         automation_dict["delay_minutes"] = int(request_data.get("delay_minutes"))
