+++ application/routes/rf_device_blueprints.py
@@ -1,15 +1,15 @@
 ################################################################################
 #
-# File:     sensor_blueprints.py
+# File:     rf_device_blueprints.py
 # Version:  0.9.0
 # Author:   Luke de Munk
-# Brief:    Flask blueprints used for sensor features.
+# Brief:    Flask blueprints used for rf_device features.
 #
 #           More information:
 #           https://github.com/LukedeMunk/zyrax-home-main-controller
 #
 ################################################################################
-from flask import Blueprint, request                                            #Import flask blueprints and requests
+from flask import Blueprint, request, session                                   #Import flask blueprints and requests
 import configuration as c                                                       #Import application configuration variables
 from DeviceManager import DeviceManager                                         #Import device manager
 from server_manager import generate_json_http_response
@@ -17,25 +17,29 @@ from logger import logi, logw, loge                                             
 import ast                                                                      #For parsing JSON within HTTP parameters
 
 dm = DeviceManager()
-sensor_bp = Blueprint("sensor_blueprints", __name__)
+rf_device_bp = Blueprint("rf_device_blueprints", __name__)
 
 ################################################################################
 #
-#   @brief  Adds a sensor to the database.
+#   @brief  Adds a RF device to the database.
 #
 ################################################################################
-@sensor_bp.route("/add_sensor", methods=["POST"])
-def add_sensor():
+@rf_device_bp.route("/add_rf_device", methods=["POST"])
+def add_rf_device():
+    if "account_id" not in session:
+        return generate_json_http_response(c.HTTP_CODE_UNAUTHORIZED)
+    
     config = {
         "name" : request.form.get("name"),
-        "model" : int(request.form.get("model")),
-        "sensor_type" : int(request.form.get("sensor_type")),
+        "model_id" : int(request.form.get("model_id")),
         "icon": request.form.get("icon"),
-        "icon_low_state": request.form.get("icon_low_state"),
         "rf_codes" : ast.literal_eval(request.form.get("rf_codes"))
     }
+
+    if "icon_low_state" in request.form:
+        config["icon_low_state"] = request.form.get("icon_low_state")
     
-    result = dm.add_sensor(config)
+    result = dm.add_rf_device(config)
     if not result[0]:
         return generate_json_http_response(c.HTTP_CODE_BAD_REQUEST, result[1])
 
@@ -43,11 +47,14 @@ def add_sensor():
 
 ################################################################################
 #
-#   @brief  Updates the specified sensor.
+#   @brief  Updates the specified RF device.
 #
 ################################################################################
-@sensor_bp.route("/update_sensor", methods=["POST"])
-def update_sensor():
+@rf_device_bp.route("/update_rf_device", methods=["POST"])
+def update_rf_device():
+    if "account_id" not in session:
+        return generate_json_http_response(c.HTTP_CODE_UNAUTHORIZED)
+    
     id = int(request.form.get("id"))
 
     config = {}
@@ -65,11 +72,14 @@ def update_sensor():
 
 ################################################################################
 #
-#   @brief  Deletes the specified sensor from the database.
+#   @brief  Deletes the specified RF device from the database.
 #
 ################################################################################
-@sensor_bp.route("/delete_sensor", methods=["POST"])
-def delete_sensor():
+@rf_device_bp.route("/delete_rf_device", methods=["POST"])
+def delete_rf_device():
+    if "account_id" not in session:
+        return generate_json_http_response(c.HTTP_CODE_UNAUTHORIZED)
+    
     id = int(request.form.get("id"))
     dm.delete_device(id)
